Add vite.config.ts based template in nx remix plugin sources
A vite.config.ts file is not emitted by the generator for the nx
remix plugin.

This commit adds a vite.config.ts based on that template, with no
further changes.

Note that the Remix application will now no longer run by default
on port 3000.

With this config file, the server does start up and can resolve
imports from the library projects in the monorepo.

However, there are console errors/warnings around stating that
'"entry.client.tsx" is outside of Vite serving allow list'.

This appears to be related to the Vite workspace detection may be
resolving the workspace directory the Remix application directory
rather than the root of the monorepo (where node_modules lives).