Update vite.config.ts to allow resolution from parent directory
This clears the previous error where the Remix application would
complain on startup stating that '"entry.client.tsx" is outside of
Vite serving allow list'.

It is not clear if this is the correct fix, or if there are any
downsides to doing this.

With this, there are no longer errors in the console when running
the application, imports to other libraries in the monorepo still
work, and hot-module-reloading due to changes in those libraries
all still works.