Upgrade nx and remix
Migrate nx to latest version, the migration deleted vite.config.ts
so this was restored.

Upgrade remix to latest version, this required npm install to use
"--legacy-peer-deps".

After nx reset, deleting nx cache directory, and executing the
following command...

    npx nx run-many -t build lint test e2e --skip-nx-cache

...everything appears to be working correctly.

It is not clear if the remix upgrade was required or not.